--- packages/react-components/src/components/FieldDescription/FieldDescription.module.scss
@@ -1,4 +1,0 @@
-.field-description {
-  display: block;
-  color: var(--content-subtle);
-}
