+++ packages/react-components/src/components/FieldDescription/FieldDescription.tsx
@@ -4,16 +4,14 @@ import cx from 'clsx';
 
 import { Text } from '../Typography';
 
-import styles from './FieldDescription.module.scss';
+import * as styles from './styles';
 
 export type FieldDescriptionProps = React.HTMLAttributes<HTMLSpanElement>;
-
-const baseClass = 'field-description';
 
 export const FieldDescription: React.FC<
   React.PropsWithChildren<FieldDescriptionProps>
 > = ({ children, className = '', ...props }) => {
-  const mergedClassNames = cx(styles[baseClass], className);
+  const mergedClassNames = cx(styles.fieldDescription, className);
 
   return (
     <Text as="span" size="sm" {...props} className={mergedClassNames}>
