+++ packages/react-components/src/components/FieldDescription/styles.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/css';
+
+import { DesignToken } from '../../foundations';
+
+export const fieldDescription = css`
+  display: block;
+  color: var(${DesignToken.ContentBasicSecondary});
+`;
