+++ packages/react-components/src/components/FieldGroup/FieldGroup.tsx
@@ -5,7 +5,7 @@ import cx from 'clsx';
 import { FieldDescription } from '../FieldDescription';
 import { FieldError } from '../FieldError';
 
-import styles from './FieldGroup.module.scss';
+import * as styles from './styles';
 
 export interface FieldGroupProps extends React.HTMLAttributes<HTMLDivElement> {
   description?: React.ReactNode;
@@ -13,8 +13,6 @@ export interface FieldGroupProps extends React.HTMLAttributes<HTMLDivElement> {
   inline?: boolean;
   stretch?: boolean;
 }
-
-const baseClass = 'field-group';
 
 export const FieldGroup: React.FC<React.PropsWithChildren<FieldGroupProps>> = ({
   className = '',
@@ -25,10 +23,7 @@ export const FieldGroup: React.FC<React.PropsWithChildren<FieldGroupProps>> = ({
   stretch,
   ...props
 }) => {
-  const mergedClassNames = cx(styles[baseClass], className, {
-    [styles[`${baseClass}--inline`]]: inline,
-    [styles[`${baseClass}--stretched`]]: stretch,
-  });
+  const mergedClassNames = cx(styles.fieldGroup(inline, stretch), className);
 
   return (
     <div {...props} className={mergedClassNames}>
