+++ packages/react-components/src/components/FieldGroup/styles.ts
@@ -1,4 +1,6 @@
-.field-group {
+import { css } from '@emotion/css';
+
+export const fieldGroup = (inline?: boolean, stretch?: boolean) => css`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -8,18 +10,21 @@
     padding: 0 0 8px;
   }
 
-  &--inline {
-    flex-direction: row;
+  ${inline &&
+  `
+     flex-direction: row;
     margin: 0 -8px;
 
     > * {
       padding: 0 8px;
     }
-  }
+    
+  `}
 
-  &--stretched {
+  ${stretch &&
+  `
     > * {
       width: 100%;
     }
-  }
-}
+  `}
+`;
